Refactored server into more manageable files.
